move enums to own files
